Introducing custom GUI support for the object browser.
Currently the object browser is only generating a default GUI. This
default gui can now be extracted into a FXML document, which is
accounting for a mapping between identifiers and slots of the
Fmmlxdiagram.

Furthermore, this extracted FXML document may be adapted in all
foreseeable manners (except changing controller, ids or actions in an
unintended way). This might be achieved by reading it within the tool
"Glu on".

Afterwards, the created custom GUI can be imported into the object
browser to replace the by default generated GUI. The values and actions
are still correctly handled by the implemented controller class.