bug(startscripts) adapt startscript to deal with changes in PropertiesLoader
Build Routine was changed. In this process there was added an
if-statment to the UserProperties. Depending on the env-var
"XMODELER_STAGE" the user.properties file is loaded from another
location. In the setup.exe that is provided by the new build routine
this file needs to show to appdata. In the case XModeler is started with
a startup file, the propertiesfile shouldbe stored relative to the
execution path. Therefore in the scripts XMODELER_STAGE is set to dev.